# AK: Increase LittleEndianOutputBitStream's buffer size and remove loops